feat: Add gridview borders for Top Beeps
Accomplished this by adding a background color gridview, horizontalSpacing, and padding for GridView.  In this way, the background color peeks through around and in-between each grid view.  Would like to improve this design later on, but it doesn't seem easy to implement rounded corners and strokes on a grid item.