fix: Change android.ndk platformVersion to 21
Previously, app was crashing on start when installed on API level 23 device.  It appears that arm64 version .so file was not installed with previous platformVersion.  Now app does not crash on start when installed on this device.  The arm64 version audio library was installed on the device.  There does seem to be some random crashing occuring during recording - sometimes is working, sometimes is not