feat: Add permission check for READ EXTERNAL memory
Note: Currently permission is checked in onCreate of RecordActivity.  This should probably be done in the SaveFragment, as that is where permission is actually needed to access the images on the device.  The Audio permission is currently checked when app starts up.  This might be better placed in the onCreate of RecordActivity.  App no longer crashes when selecting an image on M device in SaveFragment