feat: Add background to widget list items
Note: This background draws a divider line at the top of the list item. I'd like to draw a divider line at the top for all the list items except the last, where I'd like to draw a divider at the bottom of the list item. Perhaps, can do this programmatically for last board laster on