feat: Implement createWav() and make changes throughout to write correct audio
Note: Slomo effect used 'tempo' = 0.4 in the player. It seems that this is also changing the pitch in the player. However, when using the SuperpoweredTimeStretching class for offline processing, there is not a pitchshift when the 'rate' is set at 0.4. Maybe a newer version of Superpowered has fixed this bug. Also, not sure of the difference between 'tempo' and 'rate'. For now, I've set tempo/rate to 0.5 so that there is not a difference between offline and player processing