fix: Intermittent share issue fixed, but mp3 files not deleted
Previously, in onActivityResult, I was deleting the mp3 file. The sharing function was intermittently working because the sending was not necessarily done before the call to delete. When I remove the delete call, the sharing is consistently working. However, the created mp3 is not left hanging around. Maybe it's ok - mp3 files are pretty small. Probably need to work on the design of this function more though