feat: Implement ffmpeg creation and sharing using fixed image, audio
Currently the code is located in MainActivity. It takes a few seconds to create the vid, would probably take quite a few seconds with a longer 10-15 second audio clip

Note: ffmpeg does not work unless the image width and height are multiples of 2. I need to go back to my image creation code to update this